Generates working ROM
In the last commit, I didn't modify the ROM to load the logo from its own data
section instead of the cartridge.  The color comparison method introduced last
commit was also flawed and has been reverted back to its original state.
The program now generates a fully functional ROM which has been successfully
used as a drop-in replacement for a Game Boy emulator.